+++ editor/src/lib.rs
@@ -11,5 +11,6 @@ pub mod application;
 pub mod consts;
 pub mod dispatcher;
 pub mod messages;
+pub mod node_graph_executor;
 pub mod test_utils;
 pub mod utility_traits;
