+++ node-graph/graph-craft/src/document.rs
@@ -9,7 +9,6 @@ use rand_chacha::{
 	ChaCha20Rng,
 };
 use std::collections::{HashMap, HashSet};
-use std::net;
 use std::sync::Mutex;
 
 pub mod value;
